moved non-class functions to a separate file
