+++ code/components/jomjol_mqtt/CMakeLists.txt
@@ -2,6 +2,4 @@ FILE(GLOB_RECURSE app_sources ${CMAKE_CURRENT_SOURCE_DIR}/*.*)
 
 idf_component_register(SRCS ${app_sources}
                     INCLUDE_DIRS "."
-                    REQUIRES tflite-lib mqtt jomjol_logfile)
-
-
+                    REQUIRES tflite-lib mqtt  jomjol_tfliteclass jomjol_helper jomjol_mqtt jomjol_wlan)
